Update training function and callbacks for mixed precision support
- Added precision parameter to the training function to enable mixed precision training with bf16.
- Configured float32 matmul precision for improved performance on Ampere+ GPUs.
- Ensured tensor operations in SampleGridCallback are performed in float format for consistency.

+++ src/generative_models/training/callbacks.py
@@ -26,7 +26,7 @@ class SampleGridCallback(L.Callback):
             return
 
         with torch.no_grad():
-            samples = pl_module.sample(self.n_samples)
+            samples = pl_module.sample(self.n_samples).float()
         grid = make_grid(samples, nrow=4, normalize=True, value_range=(0, 1))
 
         for logger in trainer.loggers:
@@ -56,10 +56,10 @@ class SampleGridCallback(L.Callback):
             fwd = pl_module(x[: self.n_samples])
 
         originals = make_grid(
-            x[: self.n_samples], nrow=4, normalize=True, value_range=(0, 1),
+            x[: self.n_samples].float(), nrow=4, normalize=True, value_range=(0, 1),
         )
         recons = make_grid(
-            fwd["x_hat"][: self.n_samples], nrow=4, normalize=True, value_range=(0, 1),
+            fwd["x_hat"][: self.n_samples].float(), nrow=4, normalize=True, value_range=(0, 1),
         )
 
         for logger in trainer.loggers:
